Removed built-in 2x scaling given that it is not faster than pixbuf scaling. Needed to prepare automatic scaling from window size (part 2).
git-svn-id: http://svn.tilp.info/repos/tiemu@1471 4f589446-7ac1-0310-82c7-a055b08196ef

+++ tiemu/trunk/src/gui/calc/calc.h
@@ -82,7 +82,7 @@ typedef struct
 
 typedef struct
 {
-	int			t;	// twice
+	float		t;	// twice
 
 	float		x;	// x ratio
 	float		y;	// y ratio
